requirements.txt
pytorch
matplotlib
numpy
pyDOE
scipy
pickle